+++ lib/database/daos/resposta_prova.dao.dart
@@ -79,7 +79,27 @@ class RespostaProvaDao extends DatabaseAccessor<RespostasDatabase> with _$Respos
     return (delete(respostaProvaTable)..where((t) => t.sincronizado)).go();
   }
 
-  getTotalSincronizadas() {}
+  Future<int?>? getTotalSincronizadas() async {
+    final count = respostaProvaTable.sincronizado.count();
 
-  getTotalPendentes() {}
+    final query = selectOnly(respostaProvaTable)
+      ..addColumns([count])
+      ..where(respostaProvaTable.sincronizado.not());
+
+    final result = await query.get();
+
+    return result.first.read(count);
+  }
+
+  Future<int?>? getTotalPendentes() async {
+    final count = respostaProvaTable.sincronizado.count();
+
+    final query = selectOnly(respostaProvaTable)
+      ..addColumns([count])
+      ..where(respostaProvaTable.sincronizado);
+
+    final result = await query.get();
+
+    return result.first.read(count);
+  }
 }
