+++ lib/ui/views/prova/questao_revisao.view.dart
@@ -25,6 +25,7 @@ import 'package:appserap/ui/widgets/audio_player/audio_player.widget.dart';
 import 'package:appserap/ui/widgets/bases/base_state.widget.dart';
 import 'package:appserap/ui/widgets/bases/base_statefull.widget.dart';
 import 'package:appserap/ui/widgets/buttons/botao_default.widget.dart';
+import 'package:appserap/ui/widgets/status_sincronizacao/status_sincronizacao.widget.dart';
 import 'package:appserap/ui/widgets/texts/texto_default.widget.dart';
 import 'package:appserap/utils/assets.util.dart';
 import 'package:appserap/utils/file.util.dart';
@@ -129,6 +130,7 @@ class _QuestaoRevisaoViewState extends BaseStateWidget<QuestaoRevisaoView, Quest
         children: [
           TempoExecucaoWidget(provaStore: provaStore),
           _buildAudioPlayer(),
+          StatusSincronizacao(),
           Expanded(
             child: SingleChildScrollView(
               child: Padding(
